Moved Host into App Moved the .NET Host into `App.axaml.cs` and replaced the old DesignTime provider with a new one
